envs/robot_city_env/pyproject.toml
[project]
name = "robot_city_env"
version = "0.1.0"
description = "Visual multi-robot simulation environment for OpenEnv"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.20",
    "pillow>=9.0",
    "pydantic>=2.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
